requirements.txt
pyYAML
logbook
gpiozero
flask
flask_bootstrap
